requirements.txt
aiocontextvars==0.2.2
anyio==3.4.0
appnope==0.1.2
argon2-cffi==21.1.0
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
bokeh==2.3.3
certifi==2020.6.20
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorcet==3.0.0
contextvars==2.4
cycler==0.11.0
dataclasses==0.8
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
Flask==2.0.2
holoviews==1.14.6
hvplot==0.7.3
idna==3.3
immutables==0.16
importlib-metadata==4.8.2
ipykernel==5.5.6
ipython==7.16.2
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.0
json5==0.9.6
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-contrib-core==0.3.3
jupyter-core==4.9.1
jupyter-nbextensions-configurator==0.4.1
jupyter-server==1.13.0
jupyterlab==3.2.4
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.1
loguru==0.5.3
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
nbclassic==0.3.4
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.4
nltk==3.6.5
notebook==6.4.6
numpy==1.19.5
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
panel==0.12.1
param==1.12.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.12.0
prompt-toolkit==3.0.23
ptyprocess==0.7.0
pycparser==2.21
pyct==0.4.8
Pygments==2.10.0
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
pyviz-comms==2.1.0
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.1
QtPy==1.11.3
regex==2021.11.10
requests==2.26.0
scikit-learn==0.24.2
scipy==1.5.4
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tornado==6.1
tqdm==4.62.3
traitlets==4.3.3
typing_extensions==4.0.1
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.0.2
widgetsnbextension==3.5.2
zipp==3.6.0
